Rename DataDocDash to app.py
