+++ datadoc/__init__.py
@@ -1,3 +1,3 @@
-from datadoc.DataDocDash import main
+from datadoc.app import main
 
 __version__ = "0.1.1"
